+++ src/But/System/TempFile.hpp
@@ -1,5 +1,5 @@
 #pragma once
-#include <boost/filesystem/path.hpp>
+#include <filesystem>
 #include <But/System/Descriptor.hpp>
 #include <But/System/FsError.hpp>
 
@@ -13,23 +13,23 @@ class TempFile final
 public:
   BUT_DEFINE_EXCEPTION(CannotCreateTemporaryFile, FsError, "cannot crate temporary file");
 
-  explicit TempFile(boost::filesystem::path const& nameTemplate):
+  explicit TempFile(std::filesystem::path const& nameTemplate):
     file_{ initialize(nameTemplate) }
   { }
 
-  boost::filesystem::path const& path() const { return file_.path_; }
-  Descriptor const& descriptor() const { return file_.fd_; }
+  auto const& path() const { return file_.path_; }
+  auto const& descriptor() const { return file_.fd_; }
 
   void unlink();
 
 private:
   struct File
   {
-    boost::filesystem::path path_;
+    std::filesystem::path path_;
     Descriptor fd_;
   };
 
-  static File initialize(boost::filesystem::path const& tmp);
+  static File initialize(std::filesystem::path const& tmp);
 
   File file_;
 };